Bug 1672670 - move SWGL texture functions into texture.h. r=jrmuizel
Differential Revision: https://phabricator.services.mozilla.com/D94434
